gui/log: add Linux specific handler
Using the "Desktop Notifications Specification" through DBus.
